requirements.txt
# Python Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML and Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# HTTP and API
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0

# Security and Cryptography
cryptography==41.0.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Async and Background Tasks
celery==5.3.4
redis==5.0.1

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3

# Threat Intelligence
stix2==3.1.1
taxii2-client==3.0.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
